feat: add `from_numeral` function to turn Roman numerals into an int
First time using the `matches!()` macro, pretty neat!
But man, this is probably not the most elegant solution, though it makes
excellent use of Rust's standard library and pattern matching.

It takes a string, and turns it into an iterator of chunks of
characters. But the chunk size is predicated on a callback. This
callback checks if this chuck is a Roman digit that is made of multiple
parts - i.e. IV. So CDXCIV turns into an iterator over:
`[ &['C', 'D'], &['X', 'C'], &['I', 'V'] ]`, whereas VIII becomes:
`[&['V'], &['I'], &['I'], &['I']]`. It then pattern-matches the contents of
these chunks against the list of known Roman numeral digits to get the
integer value.

I also split up my existing test cases into `from_numeral` and
`to_numeral` versions. I added another case for 494, to check the
chunking logic mentioned above in particular.